backend/requirements.txt
fastapi[all]
uvicorn
motor
python-dotenv
cryptography
pydantic[email]
mangum
